fix(sign-doc): impl tryfrom instead of new and move err to own file
